requirements.txt
charset-normalizer==3.1.0
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
einops==0.6.1
huggingface-hub==0.14.1
matplotlib==3.5.3
nltk==3.8.1
numpy
packaging==23.0
pandas==1.3.5
Pillow
requests==2.29.0
scikit-learn
scipy
sentencepiece==0.1.99
tokenizers==0.13.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.65.0
transformers==4.28.1
